+++ frb_example/dart_minimal/test/minimal_test.dart
@@ -16,9 +16,4 @@ Future<void> main() async {
     print('Action: Call rust (after)');
   });
   print('Action: Configure tests (end)');
-
-  test('MyStructWithTryFromTwinNormal', () async {
-    final object = await MyStructWithTryFromTwinNormal.tryFrom(value: 'hello');
-    expect(await object.valueTwinNormal(), 'hello');
-  });
 }
