+++ frb_example/pure_dart/rust/src/api/impl_trait.rs
@@ -62,3 +62,22 @@ impl MyTraitWithSelfTwinNormal for MyImplTraitWithSelfTwinNormal {
         Self
     }
 }
+
+#[frb(opaque)]
+pub struct MyStructWithTryFromTwinNormal(String);
+
+// #2103
+impl TryFrom<String> for MyStructWithTryFromTwinNormal {
+    type Error = flutter_rust_bridge::for_generated::anyhow::Error;
+
+    #[frb]
+    fn try_from(value: String) -> flutter_rust_bridge::for_generated::anyhow::Result<Self> {
+        Ok(Self(value))
+    }
+}
+
+impl MyStructWithTryFromTwinNormal {
+    pub fn value_twin_normal(&self) -> String {
+        self.0.to_owned()
+    }
+}
